fix docs, old test references, import cycles, make frames and videos required
