+++ docs/api/devices/index.md
@@ -4,5 +4,6 @@ High-level control of external devices like miniscopes and DAQs
 
 ```{toctree}
 base
+sdcard
 stream
 ```
