--- docs/api/models/sdcard.md
@@ -1,7 +1,0 @@
-# sdcard
-
-```{eval-rst}
-.. automodule:: mio.models.sdcard
-    :members:
-    :undoc-members:
-```
