--- docs/api/models/stream.md
@@ -1,7 +1,0 @@
-# stream
-
-```{eval-rst}
-.. automodule:: mio.models.stream
-    :members:
-    :undoc-members:
-```
