+++ mio/devices/sdcard/__init__.py
@@ -1,11 +1,16 @@
 """SDCard-based miniscopes, also known as the "wire free" miniscopes"""
 
-from mio.devices.sdcard.device import SDCardDevice
+# ruff: noqa: I001
+
 from mio.devices.sdcard.headers import SDBufferHeader, SDConfig, SDHeaderPositions, SDLayout
+from mio.devices.sdcard.data import SDCardVideo, SDCardFrame
+from mio.devices.sdcard.device import SDCardDevice
 
 __all__ = [
     "SDBufferHeader",
     "SDCardDevice",
+    "SDCardFrame",
+    "SDCardVideo",
     "SDConfig",
     "SDHeaderPositions",
     "SDLayout",
