+++ mio/models/devupdate.py
@@ -2,10 +2,16 @@
 Models for device update commands.
 """
 
+import sys
 from enum import Enum
 
 import serial.tools.list_ports
 from pydantic import BaseModel, field_validator, model_validator
+
+if sys.version_info < (3, 11):
+    from typing_extensions import Self
+else:
+    from typing import Self
 
 
 class DeviceCommand(Enum):
@@ -59,12 +65,12 @@ class DevUpdateCommand(BaseModel):
     value: int
 
     @model_validator(mode="after")
-    def validate_values(cls, values: dict) -> dict:
+    def validate_values(self) -> Self:
         """
         Validate values based on key.
         """
-        key = values.key
-        value = values.value
+        key = self.key
+        value = self.value
 
         if key == UpdateKey.LED:
             assert 0 <= value <= 100, "For LED, value must be between 0 and 100"
@@ -81,7 +87,7 @@ class DevUpdateCommand(BaseModel):
             raise NotImplementedError()
         else:
             raise ValueError(f"{key} is not a valid update key," "need an instance of UpdateKey")
-        return values
+        return self
 
     @field_validator("port")
     def validate_port(cls, value: str) -> str:
