+++ mio/models/frames.py
@@ -2,21 +2,21 @@
 Pydantic models for storing frames and videos.
 """
 
+from __future__ import annotations
+
 from abc import abstractmethod
 from pathlib import Path
+from typing import TYPE_CHECKING, TypeAlias
 from typing import Annotated as A
-from typing import Literal, TypeAlias, overload
 
 import cv2
 import numpy as np
-import pandas as pd
-from numpydantic import NDArray, NDArraySchema
-from pydantic import BaseModel, Field, field_validator
+from numpydantic import NDArraySchema
+from pydantic import BaseModel, Field
 
-from mio.devices.sdcard.headers import SDBufferHeader
-from mio.logging import init_logger
+if TYPE_CHECKING:
+    pass
 
-logger = init_logger("model.frames")
 
 FrameType: TypeAlias = A[np.ndarray, NDArraySchema(("*", "*"))]
 
@@ -26,10 +26,7 @@ class BaseFrame(BaseModel):
     Pydantic model to store an image
     """
 
-    frame: FrameType | None = Field(
-        None,
-        description="Frame data, if provided.",
-    )
+    frame: FrameType
 
     @abstractmethod
     def export(self, output_path: Path | str, suffix: bool = False) -> None:
@@ -73,15 +70,10 @@ class NamedFrame(BaseFrame):
         The file name will be a concatenation of the output path and the name of the frame.
         """
         output_path = Path(output_path)
-        if self.frame is None:
-            logger.warning(f"No frame data provided for {self.name}. Skipping export.")
-            return
+
         if suffix:
             output_path = output_path.with_name(output_path.stem + f"_{self.name}")
         cv2.imwrite(str(output_path.with_suffix(".png")), self.frame)
-        logger.info(
-            f"Writing frame to {output_path}.png: {self.frame.shape[1]}x{self.frame.shape[0]}"
-        )
 
     def display(self, binary: bool = False) -> None:
         """
@@ -92,10 +84,6 @@ class NamedFrame(BaseFrame):
         binary : bool
             If True, the frame will be scaled to the full range of uint8.
         """
-        if self.frame is None:
-            logger.warning(f"No frame data provided for {self.name}. Skipping display.")
-            return
-
         frame_to_display = self.frame
         if binary:
             frame_to_display = cv2.normalize(
@@ -127,98 +115,15 @@ class NamedVideo(BaseVideo):
         """
         from mio.io import VideoWriter
 
-        if self.video is None or self.video == []:
-            logger.warning(f"No frame data provided for {self.name}. Skipping export.")
-            return
         output_path = Path(output_path)
         if suffix:
             output_path = output_path.with_name(output_path.stem + f"_{self.name}")
         if not all(isinstance(frame, np.ndarray) for frame in self.video):
             raise ValueError("Not all frames are numpy arrays.")
         writer = VideoWriter(path=output_path.with_suffix(".avi"), fps=fps, force=force)
-        logger.info(
-            f"Writing video to {output_path}.avi:"
-            f"{self.video[0].shape[1]}x{self.video[0].shape[0]}"
-        )
         try:
             for frame in self.video:
                 picture = cv2.cvtColor(frame, cv2.COLOR_GRAY2BGR)
                 writer.write_frame(picture)
         finally:
             writer.close()
-
-
-class SDCardFrame(BaseModel):
-    """
-    An individual frame from a miniscope recording
-
-    Typically returned from :meth:`.SDCardDevice.read`
-    """
-
-    frame: NDArray
-    headers: list[SDBufferHeader]
-
-    @field_validator("headers")
-    @classmethod
-    def frame_nums_must_be_equal(cls, v: list[SDBufferHeader]) -> list[SDBufferHeader] | None:
-        """
-        Each frame_number field in each header must be the same
-        (they come from the same frame!)
-        """
-
-        if v is not None and not all([header.frame_num != v[0].frame_num for header in v]):
-            raise ValueError(f"All frame numbers should be equal! Got f{[h.frame_num for h in v]}")
-        return v
-
-    @property
-    def frame_num(self) -> int | None:
-        """
-        Frame number for this set of headers, if headers are present
-        """
-        return self.headers[0].frame_num
-
-
-class SDCardVideo(BaseModel):
-    """
-    A collection of frames from a miniscope recording
-    """
-
-    frames: list[SDCardFrame]
-
-    @overload
-    def flatten_headers(self, as_dict: Literal[False]) -> list[SDBufferHeader]: ...
-
-    @overload
-    def flatten_headers(self, as_dict: Literal[True]) -> list[dict]: ...
-
-    def flatten_headers(self, as_dict: bool = False) -> list[dict] | list[SDBufferHeader]:
-        """
-        Return flat list of headers, not grouped by frame
-
-        Args:
-            as_dict (bool): If `True`, return a list of dictionaries, if `False`
-                (default), return a list of :class:`.SDBufferHeader` s.
-        """
-        h: list[dict] | list[SDBufferHeader] = []
-        for frame in self.frames:
-            headers: list[dict] | list[SDBufferHeader]
-            if as_dict:
-                headers = [header.model_dump() for header in frame.headers]
-            else:
-                headers = frame.headers
-            h.extend(headers)
-        return h
-
-    def to_df(self, what: Literal["headers"] = "headers") -> pd.DataFrame:
-        """
-        Convert frames to pandas dataframe
-
-        Arguments:
-            what ('headers'): What information from the frame to include in the df,
-                currently only 'headers' is possible
-        """
-
-        if what == "headers":
-            return pd.DataFrame(self.flatten_headers(as_dict=True))
-        else:
-            raise ValueError("Return mode not implemented!")
